refactor(api): split board-message.service.ts into services/board-message/ (api-7ds5)
