+++ apps/api/src/routes/board-messages/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { archiveBoardMessage } from "../../services/board-message.service.js";
+import { archiveBoardMessage } from "../../services/board-message/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
