+++ apps/api/src/routes/board-messages/delete.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { deleteBoardMessage } from "../../services/board-message.service.js";
+import { deleteBoardMessage } from "../../services/board-message/delete.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
