+++ apps/api/src/routes/board-messages/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getBoardMessage } from "../../services/board-message.service.js";
+import { getBoardMessage } from "../../services/board-message/queries.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
