+++ apps/api/src/routes/board-messages/list.ts
@@ -6,7 +6,7 @@ import { requireIdParam } from "../../lib/id-param.js";
 import { parsePaginationLimit } from "../../lib/pagination.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
 import { DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT } from "../../service.constants.js";
-import { listBoardMessages, parseBoardMessageQuery } from "../../services/board-message.service.js";
+import { listBoardMessages, parseBoardMessageQuery } from "../../services/board-message/queries.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
