+++ apps/api/src/routes/board-messages/pin.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { pinBoardMessage } from "../../services/board-message.service.js";
+import { pinBoardMessage } from "../../services/board-message/pin.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
