+++ apps/api/src/routes/board-messages/update.ts
@@ -7,7 +7,7 @@ import { requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { updateBoardMessage } from "../../services/board-message.service.js";
+import { updateBoardMessage } from "../../services/board-message/update.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
